Move pages into directories
